chore(moving components out of root and use Outlet.forChild())
