+++ feature-libs/pickup-in-store/components/container/pdp-pickup-options-container/pdp-pickup-options-container.module.ts
@@ -8,13 +8,22 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { CartOutlets } from '@spartacus/cart/base/root';
 import { StoreFinderModule } from '@spartacus/storefinder';
-import { OutletPosition, provideOutlet } from '@spartacus/storefront';
+import {
+  OutletModule,
+  OutletPosition,
+  provideOutlet,
+} from '@spartacus/storefront';
 import { PickupOptionsModule } from '../../presentational/index';
 
 import { PdpPickupOptionsContainerComponent } from './pdp-pickup-options-container.component';
 
 @NgModule({
-  imports: [CommonModule, PickupOptionsModule, StoreFinderModule],
+  imports: [
+    CommonModule,
+    PickupOptionsModule,
+    StoreFinderModule,
+    OutletModule.forChild(),
+  ],
   exports: [PdpPickupOptionsContainerComponent],
   declarations: [PdpPickupOptionsContainerComponent],
   providers: [
