+++ feature-libs/pickup-in-store/components/presentational/index.ts
@@ -5,4 +5,5 @@
  */
 
 export * from './pickup-info/index';
+export * from './pickup-options/index';
 export * from './store/index';
