+++ feature-libs/pickup-in-store/components/presentational/pickup-options/pickup-options.component.spec.ts
@@ -3,8 +3,8 @@ import { Component, Input } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { I18nTestingModule } from '@spartacus/core';
+import { PickupOption } from '@spartacus/pickup-in-store/root';
 import { Observable } from 'rxjs';
-import { PickupOption } from '../../../model';
 import { PickupOptionsComponent } from './pickup-options.component';
 
 describe('PickupOptionsComponent', () => {
