+++ feature-libs/pickup-in-store/components/presentational/pickup-options/pickup-options.module.ts
@@ -8,10 +8,16 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { I18nModule } from '@spartacus/core';
+import { OutletModule } from '@spartacus/storefront';
 import { PickupOptionsComponent } from './pickup-options.component';
 
 @NgModule({
-  imports: [CommonModule, I18nModule, ReactiveFormsModule],
+  imports: [
+    CommonModule,
+    I18nModule,
+    ReactiveFormsModule,
+    OutletModule.forChild(),
+  ],
   declarations: [PickupOptionsComponent],
   exports: [PickupOptionsComponent],
 })
