--- feature-libs/pickup-in-store/root/components/container/index.ts
@@ -1,9 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './cart-pickup-options-container/index';
-export * from './pdp-pickup-options-container/index';
-export * from './pickup-in-store-order-consignment/index';
