--- feature-libs/pickup-in-store/root/components/presentational/index.ts
@@ -1,7 +1,0 @@
-/*
- * SPDX-FileCopyrightText: 2023 SAP Spartacus team <spartacus-team@sap.com>
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './pickup-options/index';
